build: added makefile
Dub will rerun grammargen on every build even if the grammar has not
changed (it has no way of knowing that it hasn't changed and I can't
figure out how to give it this information). Make solves this. This will
be used on POSIX systems.

Old behaviour remains on Windows, as make is not guaranteed to be
present. Developing on windows will be slower (not slower than it used
to be, but slower than developing on POSIX is now).